rtlib: Switch to ssize_t/size_t instead of fbint/fbuint
- It's too much work to change the entire rtlib/gfxlib2 to use custom
  typedefs everywhere.
- A lot of "int" can stay "int", don't need to become 64bit, e.g. error
  codes or x/y coordinates.
- A lot of rtlib code already uses size_t, no changes needed there.
- It's good to use "native C" data types like size_t, instead of custom
  ones.

This way the amount of changes needed, especially to system-specific code
like printer or serial I/O, will be much smaller. Thus there will be much
less unnecessary changes obfuscating the diff.

+++ src/rtlib/array_lbound.c
@@ -2,7 +2,7 @@
 
 #include "fb.h"
 
-FBCALL int fb_ArrayLBound( FBARRAY *array, int dimension )
+FBCALL ssize_t fb_ArrayLBound( FBARRAY *array, ssize_t dimension )
 {
 	/* given dimension is 1-based */
 	dimension -= 1;
